backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
mysqlclient==2.1.1
python-decouple==3.8
gunicorn==20.1.0
PyJWT==2.8.0
requests==2.31.0
